+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Unreleased
 ### Added
 - Syntax highlighting for idris/idris2 code blocks in markdown files.
+- `Idris: Activate Extension` command, to manually activate when working with non-idris files.
+- Support for commands in idris2 blocks in markdown files.
 ### Changed
 ### Fixed
+- Support completions for Idris 1 .lidr files.
 
 ## 0.0.10
 ### Added